requirements.txt
# Core ML and Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tensorflow>=2.13.0
keras>=2.13.0

# Data Handling
imbalanced-learn>=0.11.0
joblib>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
dash>=2.11.0
dash-bootstrap-components>=1.4.0

# Monitoring and Logging
pyyaml>=6.0
python-dotenv>=1.0.0

# Network and Real-time Processing
scapy>=2.5.0
pyshark>=0.6

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
tqdm>=4.65.0
colorama>=0.4.6